requirements.txt
ccxt>=4.3.89
pandas>=2.2.3
ta>=0.11.0
python-dotenv>=1.0.1
